# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IV Ingestion API
  description: |
    Comprehensive API for home inspection data processing and management.
    
    ## Features
    - File upload and processing for multiple formats (PDF, DOCX, XLSX)
    - Real-time processing status updates
    - Webhook notifications for processing events
    - User management and authentication
    - Inspection data management
    - Queue monitoring and administration
    
    ## Authentication
    The API supports two authentication methods:
    - **JWT Tokens**: For user sessions and web access
    - **API Keys**: For programmatic access and integrations
    
    ## Rate Limiting
    - Free Tier: 100 requests/15min, 10 files/day
    - Pro Tier: 1000 requests/15min, 100 files/day
    - Enterprise Tier: 10000 requests/15min, 1000 files/day
    
    ## Webhooks
    Subscribe to real-time events for processing updates, inspection changes, and system notifications.
    
    ## SDKs
    Official SDKs available for:
    - JavaScript/TypeScript
    - Python
    - PHP
    
    For more information, visit our [documentation](https://docs.iv-ingestion.com).
  version: 1.0.0
  contact:
    name: IV Ingestion Support
    email: support@iv-ingestion.com
    url: https://support.iv-ingestion.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.iv-ingestion.com/v1
    description: Production server
  - url: https://staging-api.iv-ingestion.com/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      summary: User Login
      description: Authenticate user and receive JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "password123"
              rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                success: true
                data:
                  user:
                    id: "1"
                    email: "user@example.com"
                    firstName: "John"
                    lastName: "Doe"
                    role: "user"
                    isEmailVerified: true
                    createdAt: "2025-01-01T00:00:00Z"
                    updatedAt: "2025-01-01T00:00:00Z"
                  token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  refreshToken: "refresh_token_here"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      summary: Get Current User
      description: Retrieve current user profile information
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/queues:
    get:
      summary: Queue Status
      description: Retrieve processing queue status (Admin only)
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Queue status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminQueuesResponse'
        '403':
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        rememberMe:
          type: boolean
          default: false
          description: Keep user logged in
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT access token
            refreshToken:
              type: string
              description: Refresh token for token renewal
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [user, admin, inspector]
        isEmailVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
...
    AdminQueuesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              waiting:
                type: integer
              active:
                type: integer
              completed:
                type: integer
              failed:
                type: integer
              delayed:
                type: integer
              workers:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    status:
                      type: string
                      enum: [idle, working, error]
                    currentJob:
                      type: string
                    processedJobs:
                      type: integer
                    failedJobs:
                      type: integer
                    uptime:
                      type: integer
                    lastHeartbeat:
                      type: string
                      format: date-time
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...